[KEncodingProber] Clean up float constants and C-style static casts
- Remove redundant float casts of float values
- Use float literals for float constants
- Remove duplicate SHORTCUT_THRESHOLD define, already in nsCharSetProber.h
- Remove bogus '-1' confidence from JpCntx.cpp, always clamped to >= 0
  by calling probers

+++ src/probers/nsBig5Prober.cpp
@@ -59,6 +59,6 @@ float nsBig5Prober::GetConfidence(void)
 {
     float distribCf = mDistributionAnalyser.GetConfidence();
 
-    return (float)distribCf;
+    return distribCf;
 }
 }
