+++ src/probers/nsEUCKRProber.cpp
@@ -61,6 +61,6 @@ float nsEUCKRProber::GetConfidence(void)
 {
     float distribCf = mDistributionAnalyser.GetConfidence();
 
-    return (float)distribCf;
+    return distribCf;
 }
 }
